+++ lib/semver_command/test/run.t
@@ -1,0 +1,4 @@
+Comparing two versions from the command line.
+
+  $ semver-command compare 1.2.3 1.2.4
+  "1.2.3" is *less* than "1.2.4"