+++ lib/semver_command/test/test__semver_command.ml
@@ -1,4 +1,0 @@
-let%expect_test "hello" =
-  print_s Semver_command.hello_world;
-  [%expect {| "Hello, World!" |}]
-;;