+++ CHANGES.md
@@ -1,6 +1,9 @@
 ## unreleased
 
 ### Added
+
+- Simple command to compare two versions.
+
 ### Changed
 ### Fixed
 ### Removed